requirements.txt
boto3==1.14.15
pytest==5.4.3
moto==1.3.15.dev965
python-dotenv